gce/requirements.txt
flask==1.1.1
honcho==1.0.1
gunicorn==19.9.0
mysql-connector-python==8.0.18
